requirements.txt
numpy
matplotlib
tqdm
PyYAML
scikit-learn
jupyter
